AAE-21097 move CustomAuthorizationManager
