+++ activiti-cloud-service-common/activiti-cloud-services-common-security/src/main/java/org/activiti/cloud/services/common/security/jwt/JwtGrantedAuthorityConverter.java
@@ -17,7 +17,7 @@ package org.activiti.cloud.services.common.security.jwt;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import org.activiti.cloud.services.common.security.CustomAuthorizationManager;
+import org.activiti.cloud.security.authorization.CustomAuthorizationManager;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
